+++ sdk/identity/identity/CHANGELOG.md
@@ -26,6 +26,7 @@
   - Added `regionalAuthority` property to `ClientSecretCredentialOptions` and `ClientCertificateCredentialOptions`.
   - If instead of a region, `AutoDiscoverRegion` is specified as the value for `regionalAuthority`, MSAL will be used to attempt to discover the region.
   - A region can also be specified through the `AZURE_REGIONAL_AUTHORITY_NAME` environment variable.
+- All of the credentials now accept a `tenantId`, and they now work with the `GetTokenOptions` new properties: `tenantId` and `allowMultiTenantAuthentication`. If a tenant is specified, any `getToken` request that receives a `tenantId` that doesn't match the credential's `tenantId` will throw an error, unless `getToken` gets called with the `allowMultiTenantAuthentication` property set to true.
    
 ## 2.0.0-beta.3 (2021-05-12)